api/requirements.txt
django>=3.0
djangorestframework-simplejwt==4.7.2